created class StainPanel
